[Chore][Workloads] merge attention workload files into sub-family modules
workloads/attention/ 15 per-op files collapse into 4:

- gqa.py <- gqa, gqa_decode, gqa_decode_paged, gqa_prefill,
  gqa_sliding_window, gqa_sliding_window_varlen
- mha.py <- mha, mha_decode, mha_decode_paged
- deepseek.py <- deepseek_nsa, deepseek_nsa_cmp, deepseek_nsa_topk,
  deepseek_mla_decode, deepseek_dsa_decode
- mean_pooling.py unchanged

Class bodies unchanged; 23 consumer imports retargeted. Collected
test node set is byte-identical.
