+++ benchmarks/ops/attention/bench_mha_decode_paged.py
@@ -8,7 +8,7 @@ from benchmarks.benchmark_base import BenchmarkReport, ManifestBenchmark
 from benchmarks.ops.attention.manifest_params import manifest_params, mha_decode_paged_args
 from tileops.manifest import load_workloads
 from tileops.ops import MultiHeadAttentionDecodePagedWithKVCacheFwdOp
-from workloads.attention.mha_decode_paged import MhaDecodePagedTest
+from workloads.attention.mha import MhaDecodePagedTest
 
 _OP_NAME = "MultiHeadAttentionDecodePagedWithKVCacheFwdOp"
