+++ tests/ops/attention/test_deepseek_nsa.py
@@ -6,7 +6,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import NSAFwdVarlenOp
-from workloads.attention.deepseek_nsa import NsaFwdTest as _NsaFwdTestWorkload
+from workloads.attention.deepseek import NsaFwdTest as _NsaFwdTestWorkload
 
 
 class NsaFwdTest(_NsaFwdTestWorkload, TestBase):
