+++ tests/ops/attention/test_deepseek_nsa_topk.py
@@ -4,7 +4,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import NSATopkVarlenOp
-from workloads.attention.deepseek_nsa_topk import NsaTopkTest as _NsaTopkTestWorkload
+from workloads.attention.deepseek import NsaTopkTest as _NsaTopkTestWorkload
 
 
 def _nsa_topk_torch(test, q, k_cmp, lse, block_counts, block_size, scale,
