+++ tests/ops/attention/test_mha_decode_paged.py
@@ -9,7 +9,7 @@ import torch.nn.functional as F
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import MultiHeadAttentionDecodePagedWithKVCacheFwdOp
-from workloads.attention.mha_decode_paged import (
+from workloads.attention.mha import (
     MhaDecodePagedTest as _MhaDecodePagedTestWorkload,
 )
